BUG Correcao de erro valor em int para valor em double e converter valor com vírgula para double

+++ LeiloesTDSat/LeiloesTDSat/src/ProdutosDAO.java
@@ -31,7 +31,7 @@ public class ProdutosDAO {
     try {
         prep = conn.prepareStatement(sql);
         prep.setString(1, produto.getNome());
-        prep.setInt(2, produto.getValor());
+        prep.setDouble(2, produto.getValor());
         prep.setString(3, produto.getStatus());
 
         prep.executeUpdate();
@@ -58,7 +58,7 @@ public class ProdutosDAO {
 
             produto.setId(resultset.getInt("id"));
             produto.setNome(resultset.getString("nome"));
-            produto.setValor(resultset.getInt("valor"));
+            produto.setValor(resultset.getDouble("valor"));
             produto.setStatus(resultset.getString("status"));
 
             listagem.add(produto);
